Adds a Brewfile for tracking brew installs.
Using brew bundle to keep track of Homebrew dependencies so that they
can be easily reinstalled on new machines. In the future, I may want to
template this file in some way since there are some dependencies that I
only need on my work machine, but for now everything can go in the same
place for simplicity.

Install with "brew bundle install" when in the .dotfiles directory.