changes save keymap to <leader>bs
Having the save keymap as <leader>s meant that it was a short form of
the "searching" keymaps (i.e. <leader>sp for searching project). Having
this as a short form of those keymaps meant that you need to wait for
the mapping timeout to complete before the :w command was executed,
which was very annoying.

I chose <leader>bs ("buffer-save") instead of <leader>fs ("file-save") because fs
requires upper-left lateral movement on dvorak which is a motion that i
don't want to be repeating all of the time. Lower-left lateral movement
feels much more natural for me.