tmux prefix C-Space; vim write file <leader>,
Updates the tmux prefix to C-Space from C-s. This was done to limit the
usage of the right pinky; avoiding more ulnar nerve problems.

Updates vim write file to <leader>, for similar reasons. "," was chosen
as it resides in the "w" position if you were looking at the qwerty
keycodes for the dvorak layout. Alternating hands (since I always press
space with my right hand) is also a good benefit.