vim: map <CR> to write file
Why make saving more complicated than it needs to be? <CR> is useless in
normal mode as it's use is duplicated by using '+'. Here we make sure
that we only execute the :write command against a "normal" buffer, which
is signified by "" in buftype. This means that pressing <CR> in quickfix
lists (and other places) doesn't break.