nvim: mappings for adding lines above/below curr
Press <Leader>o to insert a newline below the current line while
retaining the cursor position. <Leader>O does the same but above the
current line.

One potential problem with this mapping is that it clobbers the `l'
register if it was set. I barely use registers anyway and can't imagine
using `l' so it's probably fine.