nvim: listchars configuration for tabs and trailing whitespace
I don't really want to automatically remove trailing whitespace or
convert all tabs to spaces because that could trigger large diffs which
is really annoying.
Configuring listchars to be visible (and changing the hl group so that I
can actually see it) is great because now I can run :Chomp to clean up a
file or fix the problems manually on smaller sections as needed.