nvim: telescope grep and close changes
Allows for the ability to close telescope from insert mode with a single
press of <esc>. I never used normal mode inside of the telescope prompt
anyway.

Changing search project keybinding from live_grep to grep_string with a
search prompt. live_grep had a lot of performance problems in my larger
repositories and it was really annoying to have to execute different
commands depending on the size of the repo.