zsh: updates t function to be able to move/kill win
There are too many times when I have created a new window in my current
session so that I could execute either `t' or `td'. Doing that means
that we leave an orphaned window in the previous session. These two new
options prevent this as they allow us to either move the current window
(because I have already been working in that window and want the
context) or kill the current window (because it was just a temp window).