zsh: compiling zsh scripts
Added ignore = dirty to submodules so that the .zwc files do not show up
in my git diff.
Opted to write a `compile' script to ensure that used files are easily
recompiled. I chose this option instead of committing the .zwc files
since I will have to compile the submodules anyway (I'm not going to
fork a repo just to commit .zwc files...I'm not crazy).