nvim: disable list option in go files
My intenion for the list option was so that I knew when a tab character
was in a file where I wanted expandtab. Go uses tabs by default (because
of `gofmt') and I don't want to see the tab listchars all over the place
in my go files. Therefore this conditionally turns them off.

Room for improvement:
I could have conditionally updated `listchars' so that I keep trailing
whitespace highlighted, however, `gofmt' will fix that for me
automatically so it's better to not have the noise.