zsh: adds push-line binding
push-line allows you to suspend the editing of a command line to run a
different command. when that command completes then your partially
edited command line is replaced.

i think this will replace my usage of commenting out the current line to
execute a command and then go back to it with the history.