nvim: reenable live_grep, move string grep to <SPC>sP
live_grep is just so much nicer than string grep so I'm going back to
making that the default. If a particular project has problems with
live_grep I'll use string grep for that case.

Also fixes a deprecation bug where telescope removed the `shorten_path`
option for the use of a `path_display` option.