fixes custom rg script usage with telescope
Some relatively recent version of telescope broke my custom ripgrep
integration with live_grep. I'm not quite sure why, but it seems to be
related to having the wrapper script as a zsh function.

Moving the wrapper script to a standalone bash script that is called by
telescope fixed the problem. Now "rg" is an alias to this script instead
of a custom zsh function all works the same (except now there is an
extra bash process in the middle :shrug:).