adds git snapshot alias
This alias pushes the stashes the current changes and then immediately
reapplies them. This is useful because it gives you a quick way to
create a checkpoint to roll back without creating a commit.

It's also useful because changes in your working index are vulnerable to
mistakes. Running 'snapshot' will ensure that blobs are created for all
changes in the current index, so these changes will be visible in the
reflog in case you do something crazy. :)