switch to flakey-profile from home-manager
I was only using home-manager to install packages, I didn't like all of
the other bits that it did. flakey-profile is a much simpler
alternative, and it's just a flake.

To switch first uninstall home-manager:

    home-manager uninstall

then when you need to add a new package it's done from the dotfiles
checkout:

    cd ~/.dotfiles/nix-profile/
    nix run .#profile.switch