bump nix/nvim packages
Upgraded to nvim 0.10.0.
Some small improvements to nvim setup to pair with it. Certain
keybindings are now default so they no longer need to be explicitly set.

Had to explicitly install fzf 0.53.0 (only 0.52.1 is available upstream)
as the --no-tmux option is being provided from... somewhere (it's not in
the source code of fzf.vim) and that's a 0.53.0 feature only. Can remove
override once 0.53.0 ships in nixpkgs.

TODO: My colorscheme is now a bit broken in nvim. My colorscheme is now
a bit broken even with the old versions of plugins so I guess the
upgrade broke it?